mingw: optionally enable wsl compability file mode bits
In the Windows Subsystem for Linux (WSL) version 2, we can set the mode
bits on an NTFS volume provided that WSL has enabled metadata mounting
NTFS volumes, for example:

$ chmod 0755 /mnt/d/test/a.sh

In order to facilitate better collaboration between the Windows
version of Git and the WSL version of Git, we can make the Windows
version of Git also support reading and writing NTFS file modes
in a manner compatible with WSL.

There are several ways to enable metadata loading for NTFS volumes
in WSL, one of which is to modify /etc/wsl.conf by adding:

```
[automount]
enabled = true
options = "metadata,umask=027,fmask=117"
```

And reboot.

It's important to note that this modification is compatible with,
but does not depend on WSL. The helper functions in this commit
can operate independently and function normally on devices
where WSL is not installed or properly configured.

This feature is only enabled when `core.wslcompat` is set to true.

Signed-off-by: xungeng li <xungeng@gmail.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>