ci: upgrade to using macos-13
It is only a matter of time until the `macos-12` pool is going away, and
we may even get a performance boost by switching over to `macos-13`.

Since the `macos-13` runners no longer include Python2, we also drop
specifically testing with Python2 and switch uniformly to Python3, see
https://github.com/actions/runner-images/blob/HEAD/images/macos/macos-13-Readme.md

Also, on macOS 13, Homebrew seems to install a `gcc@9` package that no
longer has `unistd.h`, only `ssp/unistd.h`, and hence builds would
fail with:

    In file included from base85.c:1:
    git-compat-util.h:223:10: fatal error: unistd.h: No such file or directory
      223 | #include <unistd.h>
          |          ^~~~~~~~~~
    compilation terminated.

The reason why we install GCC v9.x is historical, and back in the days
it was because it was the _newest_ version available via Homebrew:
176441bfb58 (ci: build Git with GCC 9 in the 'osx-gcc' build job,
2019-11-27).

To reinstate the spirit of that commit _and_ to fix that build failure,
let's switch to the now-newest GCC version: v13.x.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>